requirements-ml.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
easyocr
tensorflow-cpu>=2.16,<2.19
flask
gunicorn
Pillow
numpy<1.27
